Broke out views into their own files
