+++ un_project/un_app/views/__init__.py
@@ -1,0 +1,9 @@
+from .building_list_view import *
+from .company_balace_sheet_view import *
+from .evaluate_buildings_view import *
+from .evaluate_items_view import *
+from .general_building_info_view import *
+from .item_counter_view import *
+from .misc_small_views import *
+from .nation_balace_sheet_view import *
+from .player_list_view import *
