+++ un_project/un_app/views/player_list_view.py
@@ -1,0 +1,9 @@
+from django.shortcuts import render
+from django.db.models import Count
+from ..models import Player
+
+def player_list(request):
+    players = Player.objects.select_related('nation').annotate(
+        num_buildings_built=Count('main_builds')
+    ).order_by('-num_buildings_built')  # Sort by the number of buildings built in descending order
+    return render(request, 'player_list.html', {'players': players})
